requirements.txt
schwabdev>=2.5.0
dash>=2.9.0
plotly>=5.13.0
pandas>=1.5.3
numpy>=1.24.0
python-dotenv>=1.0.0
streamlit>=1.22.0
